+++ homepage/package.json
@@ -17,7 +17,7 @@
     "react-scripts": "^5.0.1"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "set NODE_ENV = dev&&react-scripts start",
     "build": "react-scripts build"
   },
   "eslintConfig": {
