--- homepage/src/components/Intro/NotFound.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-
-export default function NotFound() {
-    return (
-        <div>
-            404 NOT FOUND 😆
-        </div>
-    );
-}
-
