+++ homepage/src/components/Storage/Storage.jsx
@@ -1,58 +1,41 @@
-import axios from "axios";
-import React, { useEffect } from "react";
-import { useCookies } from "react-cookie";
-import { useNavigate } from "react-router-dom";
+import React, { useEffect, useState } from "react";
+import Sidebar from "./Sidebar";
+import MyComponent from "./MyComponent";
+import { BrowserRouter } from "react-router-dom";
 
 export default function Storage() {
-  const go = useNavigate();
-  const [cookies, removeCookie] = useCookies("accessToken");
+  const [data, setData] = useState(null);
 
   useEffect(() => {
-    axios
-      .get("http://localhost:8080/api/auth", {
-        headers: {
-          "x-auth-token": cookies.accessToken,
-        },
-      })
-      .then((res) => {
-        console.log(res);
-      })
-      .catch((error) => {
-        console.log(error);
-        removeCookie("accessToken");
-        alert("로그인이 필요합니다.")
-        go("/");
-      });
+    fetchData(); // App.js가 마운트될 때 데이터를 가져오도록 useEffect 훅 사용
   }, []);
-  const logout = () => {
-    removeCookie("accessToken");
-    go("/");
+
+  const fetchData = async () => {
+    try {
+      const response = await fetch("http://localhost:8080/keyWordByDate", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json", // JSON 형식으로 요청을 보내기 위해 Content-Type 헤더를 설정합니다.
+        },
+        body: JSON.stringify({ userId: "cofla" }), // userId를 포함한 객체를 JSON 문자열로 변환하여 본문에 포함시킵니다.
+      });
+      const data = await response.json(); // 응답 데이터를 JSON 형식으로 변환합니다.
+      setData(data); // 가져온 데이터를 상태로 설정합니다.
+    } catch (error) {
+      console.log("Error getting data:", error);
+    }
   };
+
   return (
-    <>
-      <div className="flex p-10 space-x-2 justify-between bg-orange-400">
-        <p className="text-3xl font-bold">Myname is Storage 🫠</p>
-        <button
-          onClick={logout}
-          className="bg-white text-orange-400 hover:bg-oragne-700 focus:ring-4 focus:outline-none focus:ring-orange-300 font-semibold rounded-lg text-sm px-5 py-2 dark:bg-orange-600 dark:hover:bg-orange-700 dark:focus:ring-orange-800"
-        >
-          로그아웃
-        </button>
+    <div style={{ display: "flex" }}>
+      <div style={{ border: "1px solid black", padding: "10px" }}>
+        <Sidebar data={data}/>
       </div>
-      <div className="flex mt-2 pl-2 space-x-2 justify-center">
-        <button
-          className=" text-white bg-orange-400 hover:bg-oragne-700 focus:ring-4 focus:outline-none focus:ring-orange-300 font-semibold rounded-lg text-sm px-5 py-2.5 dark:bg-orange-600 dark:hover:bg-orange-700 dark:focus:ring-orange-800"
-          onClick={() => go("/")}
-        >
-          go to intro
-        </button>
-        <button
-          className="text-white bg-orange-400 hover:bg-oragne-700 focus:ring-4 focus:outline-none focus:ring-orange-300 font-semibold rounded-lg text-sm px-5 py-2.5 dark:bg-orange-600 dark:hover:bg-orange-700 dark:focus:ring-orange-800"
-          onClick={() => go("/main")}
-        >
-          go to main
-        </button>
+      <div>
+        <BrowserRouter>
+          <MyComponent scrapdata={data} /> {/* 데이터를 MyComponent에 전달합니다 */}
+        </BrowserRouter>
       </div>
-    </>
+    </div>
   );
 }
