+++ homepage/src/index.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Storage from "./Storage";
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
-    <Storage />
+    <App />
   </React.StrictMode>
 );
